requirements.txt
Flask
python-dotenv
openai
pandas
PyPDF2
werkzeug

